[Fixes #79] NPE hides outer exception

+++ src/main/java/org/mapfish/print/servlet/MapPrinterServlet.java
@@ -201,7 +201,9 @@ public class MapPrinterServlet extends BaseMapServlet {
                 return;
             }
         } catch (Throwable e) {
-            deleteFile(tempFileMetadata.tempFile);
+            if(tempFileMetadata != null) {
+                deleteFile(tempFileMetadata.tempFile);
+            }
             error(httpServletResponse, e);
             return;
         }
